Update NixOS flake reference to most recent nixos-23.05
We reference the flake by git branch name rather than a fixed git
commit hash. This makes it clearer which nixos channel I'm following,
and the lockfile provides the hash for the specific commit anyway. (In
this case, I'm updating from nixos-22.11 to nixos-23.05.)

In addition to changing the format used to reference the flake from a
string to an attribute set. I believe I copy/pasted the original flake
reference from my list of packages (see .config/nixpkgs/packages.nix),
and because the builtins.getFlake function only accepts a string, and
I wanted to reference a fixed commit, I needed to use a URL with that
specific commit. I would have preferred to use an attribute set, but
that does not (yet?) seem to be supported.