Show all transient menu options by default
The transient library in emacs is used to show keyboard driven
menus. Each menu item has a "level", and items whose level is greater
than the current level are not shown. The maximum level is 7, so we
max out the default level to ensure that we always see all possible
options.

As far as I know, this is only used by magit.