Add elm overlay to fix obnoxious error message in "elm repl"
I recently discovered that "elm repl" produces the following obnoxious
error message on startup as well as with every key press:

elm: freeHaskellFunctionPtr: not for me, guv! 0x7f2815d1efe8

The _only_ result on Google for "freeHaskellFunctionPtr" is this
nixpkgs issue [1], which explains the problem is a result of
bad/outdated Haskell dependencies.

This issue spawned a PR which has been merged to nixpkgs master, but
hasn't made its way into any release channels (that I can see). I
don't want to update all my packages just to fix this, and there were
enough changes that targeting them specifically in an overlay seemed
like it would take more time than it was worth. I instead copy/pasted
the contents of pkgs/development/compilers/elm to my overlay and use
that as the definition of elmPackages.

This overlay can go away once I update to a version of nixpkgs which
includes these changes.

1. https://github.com/NixOS/nixpkgs/issues/277156